refactor(Application): change location of mapping and DTO
